Implemented inserting facts in KieSession using Kafka queuing system.
Implemented Fact insertion in Kie Sessions Using Kafka Consumers,
It reproduces the behavior of hacep-core-camel, instead of pushing
the events in the datagrid then using a post hook to insert the fact
in the KieSession. It consumes the data from a Topic (Route in AMQ)
and insert it in KieSession.